% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{check_trektellen_data}
\alias{check_trektellen_data}
\title{Check formatted Trektellen data}
\usage{
check_trektellen_data(data)
}
\arguments{
\item{data}{Dataframe with BRC formatted data}
}
\value{
Dataframe with checked Trektellen data. Rows are flagged for
several common data issues in the 'check' column.
}
\description{
Check formatted Trektellen data
}
