% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trektellen.R
\name{extract_count_period}
\alias{extract_count_period}
\title{Extract count period from Trektellen website}
\usage{
extract_count_period(station, date_str)
}
\arguments{
\item{station}{Station ID or name}

\item{date_str}{Date string}
}
\value{
NULL if failed, otherwise a list with start and end times
}
\description{
Extract count period from Trektellen website
}
