% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_season_params}
\alias{get_season_params}
\title{Get correct season parameters based on a year as input}
\usage{
get_season_params(yyyy)
}
\arguments{
\item{yyyy}{Year (integer) in YYYY}
}
\description{
Get correct season parameters based on a year as input
}
