% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_raw_trektellen_data}
\alias{preprocess_raw_trektellen_data}
\title{Preprocess RAW Trektellen CSV to the standardised BRC form}
\usage{
preprocess_raw_trektellen_data(csv_path, date_str = NULL)
}
\arguments{
\item{csv_path}{Path to .csv file with RAW Trektellen counts}

\item{date_str}{Date string in YYYY-MM-DD format to filter data from RAW counts}
}
\value{
a pre-processed dataframe which follows BRC data formatting
}
\description{
Preprocess RAW Trektellen CSV to the standardised BRC form
}
