% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_day}
\alias{process_day}
\title{Process data from a day}
\usage{
process_day(day_str, tt_file, output_folder)
}
\arguments{
\item{day_str}{Date string in YYYY-MM-DD format}

\item{tt_file}{Path to Trektellen .csv file}

\item{output_folder}{Path to output directory where raw/, inprogress/,
inprogress-backup/ and clean/ folders sit}
}
\value{
This function is used for the side-effect of processing a day's data
}
\description{
Process data from a day
}
