% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_today}
\alias{process_today}
\title{Process data from today}
\usage{
process_today(tt_file, output_folder)
}
\arguments{
\item{tt_file}{Path to Trektellen .csv file}

\item{output_folder}{Path to output directory where raw/, inprogress/,
inprogress-backup/ and clean/ folders sit}
}
\value{
This function is used for the side-effect of processing today's data
}
\description{
A wrapper function around \code{process_day} that defaults to today
}
