% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_data}
\alias{write_data}
\title{Write data file to .xlsx format}
\usage{
write_data(data, folder, split_by_station = TRUE, overwrite = FALSE)
}
\arguments{
\item{data}{Dataframe}

\item{folder}{Output folder}

\item{split_by_station}{Boolean; If TRUE split by station}

\item{overwrite}{Boolean; FALSE by default.}
}
\value{
This function is used for the side-effect of writing a file
}
\description{
Write data file to .xlsx format
}
